backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
filetype==1.2.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.24
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pgvector==0.4.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic_core==2.33.1
PyPDF2==3.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
zstandard==0.23.0
